requirements.txt
python-dotenv==1.0.0
openai==0.28.0
azure-ai-textanalytics==5.3.0